fix: fix unresolved dependencies in conftest when running colvision tests
