+++ tests/conftest.py
@@ -4,8 +4,6 @@ from typing import Dict, List
 import pytest
 from langchain_milvus.utils.sparse import BaseSparseEmbedding
 
-from mmore.privacy.agents.base import clear_llm_cache
-from mmore.privacy.agents.registry import tool_registry
 from mmore.type import MultimodalSample
 
 
@@ -83,19 +81,3 @@ def pytest_collection_modifyitems(config, items):
         for item in items:
             if "gpu" in item.keywords:
                 item.add_marker(skip_gpu)
-
-
-@pytest.fixture
-def isolate_llm_cache():
-    clear_llm_cache()
-    yield
-    clear_llm_cache()
-
-
-@pytest.fixture
-def isolated_tool_registry():
-    snapshot = dict(tool_registry)
-    tool_registry.clear()
-    yield
-    tool_registry.clear()
-    tool_registry.update(snapshot)
